refactor: useScheduleGlance hook
